// src/enums/negotiation-trait.ts
export enum NegotiationTrait {
	Benevolence = 'Benevolence',
	Discovery = 'Discovery',
	Freedom = 'Freedom',
	Greed = 'Greed',
	HigherAuthority = 'Higher Authority',
	Justice = 'Justice',
	Legacy = 'Legacy',
	Peace = 'Peace',
	Power = 'Power',
	Protection = 'Protection',
	Revelry = 'Revelry',
	Vengeance = 'Vengeance'
}